logging

+++ SecureSend.Application/Commands/Handlers/DeleteSecureUploadHandler.cs
@@ -18,6 +18,7 @@ namespace SecureSend.Application.Commands.Handlers
         public async Task Handle(DeleteSecureUpload request, CancellationToken cancellationToken)
         {
             var persisted = await _repository.GetAsync(request.id ,true);
+            if (persisted is null) throw new UploadDoesNotExistException(request.id);
             _fileService.RemoveUpload(request.id);
             await _repository.DeleteAsync(persisted);
         }
